v6.0.317.1 is out.
Started with a version change, but included a note about how an install can be customized.  I hate when these programs take over everything, and it took me awhile to figure out how to turn some of those options off.  Including the information is useful to others.  I'm not sure "release notes" is the best location, but I don't know a better one.